[PC-1235] Fix: FCM 토큰 등록 로직 개선
- FCM 토큰 등록 시 Access Token 확인 로작 추가
- FCM 토큰 서버 등록 책임을 AppDelegate로 이동
