+++ Utility/PCFoundationExtension/Sources/Notification.Name+.swift
@@ -10,4 +10,5 @@ import Foundation
 // MARK: - Push Notifications
 public extension Notification.Name {
   static let deepLinkHome = Notification.Name("DeepLinkHome")
+  static let fcmToken = Notification.Name("FCMToken")
 }
